Update Documentation and squash bugs
Fixed a few bugs in BasicMovement.java where the lift was moving in the opposite direction and was not taking account of the limit sensors.
Fixed CalibrationColorSensor.java's dependency on the Sensor class (it should be independent)
Add more information to TeamCode README and remove ftc_app stock README. Will update later.